Connor robertson tuesday

+++ app/src/reduxMiddleware/authenticate.js
@@ -1,10 +1,18 @@
-import { SUCCESS_LOGIN, SUCCESS_SIGN_UP } from "../actions/actionTypes";
+import { SUCCESS_LOGIN, SUCCESS_SIGN_UP, LOG_OUT, SET_AUTH } from "../actions/actionTypes";
+
 const authenticate = store => next => action => {
   if (action.type === SUCCESS_LOGIN || action.type === SUCCESS_SIGN_UP) {
-    localStorage.setItem("tokken", action.payload.token);
+    localStorage.setItem("token", action.payload.token);
+    localStorage.setItem("user", JSON.stringify(action.payload));
   }
-  if (action.payload !== undefined && action.payload.status === 401) {
+
+  if (
+    action.payload !== undefined &&
+    action.payload.response !== undefined &&
+    action.payload.response.status === 401
+  ) {
     localStorage.clear();
+    action.type = LOG_OUT;
   }
   next(action);
 };
